java 학습 내용 업로드

+++ java_bible/src/IntEx1.java
@@ -2,7 +2,8 @@
 public class IntEx1 {
 
 	public static void main(String[] args) {
-		byte var1 = 128; // 범위를 벗어나서 오류가 발생
+		// 범위를 벗어나서 오류가 발생 byte var1 = 128; 
+		int var1 = 128; 
 		System.out.println("var1의 값:"+var1);
 	}
